Handle NS and NSE server wait timeouts cleanly
gRPC server-side streaming calls should be closed by returning nil so
we need to figure out what the timeout Error looks like and return nil
when it happens. In this case the Error has an "unknown" code so we
use the error message. When it happens we return nil and the
server-side streaming call is closed cleanly.

Signed-off-by: Toby Cabot <toby@acnodal.io>